api/backend/python/xl-client/pyproject.toml
[tool.poetry]
name = "dial_xl"
version = "0.1.2"
description = ""
authors = ["Alexey Klimov <alexey_klimov@epam.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.1"
protobuf = "3.20.*"
aiohttp = "^3.9.5"
ruamel-yaml = "^0.18.6"

[tool.poetry.group.dev.dependencies]
nox = "^2023.4.22"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"

[tool.poetry.group.lint.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
isort = "^5.12.0"
pyright = "^1.1.324"
autoflake = "^2.2.0"

[tool.pyright]
typeCheckingMode = "basic"
reportUnusedVariable = "error"
reportIncompatibleMethodOverride = "error"
exclude = [
    ".git",
    ".venv",
    ".nox",
    "**/.pytest_cache",
    "**/__pycache__",
    "**/*_pb2.py",
]

[tool.black]
line-length = 80
exclude = '''
/(
    \.git
  | \.venv
  | \.nox
  | \.pytest_cache
  | __pycache__
)/
'''

[tool.isort]
line_length = 80
profile = "black"

[tool.autoflake]
ignore_init_module_imports = true
remove_all_unused_imports = true
in_place = true
recursive = true
quiet = true
exclude = [
    '\.git',
    '\.venv',
    '\.nox',
    '\.pytest_cache',
    '__pycache__',
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
